# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "NyanPass 高性能中转面板 API 文档",
        "title": "NyanPass API",
        "termsOfService": "https://nyanpass.io/terms",
        "contact": {
            "name": "API Support",
            "url": "https://nyanpass.io/support",
            "email": "support@nyanpass.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/admin/announcements": {
            "get": {
                "tags": [
                    "Admin/Announcement"
                ],
                "summary": "获取所有公告（包含草稿）",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Announcement"
                ],
                "summary": "创建公告（管理员）",
                "parameters": [
                    {
                        "description": "公告信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.CreateAnnouncementRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/announcements/{id}": {
            "get": {
                "tags": [
                    "Admin/Announcement"
                ],
                "summary": "获取公告详情（管理员）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "公告ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Announcement"
                ],
                "summary": "更新公告（管理员）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "公告ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "公告信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.UpdateAnnouncementRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admin/Announcement"
                ],
                "summary": "删除公告（管理员）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "公告ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/announcements/{id}/offline": {
            "post": {
                "tags": [
                    "Admin/Announcement"
                ],
                "summary": "下线公告（管理员）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "公告ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/announcements/{id}/publish": {
            "post": {
                "tags": [
                    "Admin/Announcement"
                ],
                "summary": "发布公告（管理员）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "公告ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/auth/login": {
            "post": {
                "description": "使用邮箱和密码登录，获取 JWT Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.LoginResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "登录失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "注册新用户账户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "注册信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "注册失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.User": {
            "type": "object",
            "properties": {
                "balance": {
                    "description": "资产信息",
                    "type": "number"
                },
                "commission": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "download": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "expired_at": {
                    "type": "string"
                },
                "group_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "invite_code": {
                    "description": "邀请系统",
                    "type": "string"
                },
                "invited_by": {
                    "type": "integer"
                },
                "is_admin": {
                    "type": "boolean"
                },
                "last_login_at": {
                    "description": "系统信息",
                    "type": "string"
                },
                "last_login_ip": {
                    "type": "string"
                },
                "status": {
                    "description": "状态与权限",
                    "type": "integer"
                },
                "telegram_id": {
                    "description": "第三方绑定",
                    "type": "integer"
                },
                "transfer_enable": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "upload": {
                    "description": "流量控制 (单位: Bytes)",
                    "type": "integer"
                },
                "uuid": {
                    "description": "订阅唯一标识 (SQLite兼容)",
                    "type": "string"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据"
                },
                "msg": {
                    "description": "提示信息",
                    "type": "string"
                }
            }
        },
...
        "service.CreateAnnouncementRequest": {
            "type": "object",
            "required": [
                "content",
                "title"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "expired_at": {
                    "type": "string"
                },
                "popup": {
                    "type": "boolean"
                },
                "priority": {
                    "type": "integer"
                },
                "published_at": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "tags": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "service.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "service.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                }
            }
        },
        "service.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "invite_code": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                }
            }
        },
...
        "service.UpdateAnnouncementRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "expired_at": {
                    "type": "string"
                },
                "popup": {
                    "type": "boolean"
                },
                "priority": {
                    "type": "integer"
                },
                "published_at": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "tags": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Token 认证，格式：Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}